pyproject.toml
[tool.poetry]
name = "lesson-drf"
version = "0.1.0"
description = ""
authors = ["Mark Pochivalov <pochivalov.mark@gmail.com>"]
readme = "README.md"
packages = [{include = "lesson_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
psycopg2 = "^2.9.7"
djangorestframework = "^3.14.0"
pillow = "^10.0.0"
python-dotenv = "^1.0.0"
django-filter = "^23.2"
docutils = "^0.20.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
